[CHORE] #11 - 빌드 세팅 변경
